Fix settings menus: cvar read bug, enum labels, dead bindings
The root cause of all slider/enum display issues was a float/double
mismatch: Cvar_VariableValue() returns double but the extern "C"
declaration said float, causing garbage reads on x86_64.

Also:
- Add BindFunc-based _label bindings for enum display (e.g. "High"
  instead of raw int 0)
- Remap non-existent cvars: gl_texturemode->vid_filter,
  hud_style->scr_style; remove r_shadows, host_fastload, hud_detail
- Fix r_scale from percentages to engine multipliers (1, 2, 4)
- Remove duplicate volume/sfxvolume binding and Effects Volume UI row
- Make ambient_level CVAR_ARCHIVE so it persists across restarts

+++ rmlui/internal/quake_cvar_provider.cpp
@@ -16,7 +16,7 @@ extern "C" {
     cvar_t* Cvar_FindVar(const char* var_name);
     void Cvar_SetValue(const char* var_name, float value);
     void Cvar_Set(const char* var_name, const char* value);
-    float Cvar_VariableValue(const char* var_name);
+    double Cvar_VariableValue(const char* var_name);
     const char* Cvar_VariableString(const char* var_name);
 }
